Renovate: Repurpose into "CrateDB Ecosystem Catalog"
After all the tutorials have been refactored into the CrateDB Guide,
the enumeration of catalog items became a bit of a lost place.

This improvement concludes the renovation on this end, by effectively
repurposing it into a ecosystem software catalog/gallery, similar to how
others are running them.

Other than this, the patch also adds concise navigation elements to the
top of each page, in order to add gravity towards the tutorial items.